seperate admins data

+++ app/Http/Controllers/Admin/Bank/BankController.php
@@ -12,7 +12,9 @@ use App\Models\Banks;
 
 class BankController extends Controller
 {
-    public function __construct(private BankService $bankService) {}
+    public function __construct(private BankService $bankService)
+    {
+    }
 
     public function index()
     {
@@ -43,12 +45,20 @@ class BankController extends Controller
 
     public function edit(Banks $bank)
     {
+        if (!$bank || $bank->admin_id != getAdminIdOrCreatedBy()) {
+            return back()
+                ->with('Error', __('admin.not_found_data'));
+        }
         $status = ActivationStatusEnum::cases();
         return view('dashboard.pages.bank.edit', compact('status', 'bank'));
     }
 
     public function update(UpdateBankRequest $request, Banks $bank)
     {
+        if (!$bank || $bank->admin_id != getAdminIdOrCreatedBy()) {
+            return back()
+                ->with('Error', __('admin.not_found_data'));
+        }
         $this->bankService->updateBank($request, $bank);
         return back()
             ->with("Success", __('admin.updated_successfully'));
@@ -56,6 +66,10 @@ class BankController extends Controller
 
     public function updateStatus(Banks $bank)
     {
+        if (!$bank || $bank->admin_id != getAdminIdOrCreatedBy()) {
+            return back()
+                ->with('Error', __('admin.not_found_data'));
+        }
         $this->bankService->updateBankStatus($bank);
         return back()
             ->with("Success", __('admin.updated_successfully'));
@@ -63,6 +77,10 @@ class BankController extends Controller
 
     public function delete(Banks $bank)
     {
+        if (!$bank || $bank->admin_id != getAdminIdOrCreatedBy()) {
+            return back()
+                ->with('Error', __('admin.not_found_data'));
+        }
         $this->bankService->deleteBank($bank);
         return back()
             ->with("Success", __('admin.deleted_successfully'));
